+++ R/outfun_core.R
@@ -81,6 +81,10 @@ compute_stats <- function(x, variable, strata = get_default_strata(), stats = c(
     msg = "variable can't be used as a stratification column name"
   )
 
+  # Preserve original order of stratification variables
+  character_strata_cols <- get_character_cols_only(x = x, cols = strata_cols)
+  x <- preserve_column_levels(x = x, cols = character_strata_cols)
+
   # Filter data
   x_filtered <- filter_output_on_strata(x = x, strata = strata)
 
@@ -116,6 +120,9 @@ compute_stats <- function(x, variable, strata = get_default_strata(), stats = c(
 
   # Pivot longer into the final format
   res <- metrics_pivot_longer(x = res, cols = stats)
+
+  # Stratification variables back to character vectors
+  res <- remove_column_levels(x = res, cols = character_strata_cols)
 
   # Variable before TIME
   res <- res |>