+++ R/simulation_engine.R
@@ -18,8 +18,7 @@ setClass(
 #' @export
 setClass(
   "rxode_engine",
-  representation(
-  ),
+  representation(),
   contains = "simulation_engine"
 )